requirements.txt
cassandra-driver==3.25.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
kafka-python==2.0.2
pandas==1.4.2
numpy==1.22.4
python-dateutil==2.8.2
pytz==2022.1
requests==2.25.1
six==1.15.0
sseclient==0.0.27
urllib3==1.26.5
